make simulator var a const, move `onContentScriptConnected` to startup
