+++ app/ts/background/background-startup.ts
@@ -1,30 +1,141 @@
 import 'webextension-polyfill'
 import { getMakeMeRich, getSettings } from './settings.js'
-import { getPrependTrasactions, onContentScriptConnected, popupMessageHandler, updateSimulationState } from './background.js'
-import { updateExtensionBadge } from './iconHandler.js'
-import { clearTabStates, getSimulationResults, removeTabState, setRpcConnectionStatus } from './storageVariables.js'
+import { gateKeepRequestBehindAccessDialog, getPrependTrasactions, handleContentScriptMessage, popupMessageHandler, refuseAccess, updateSimulationState } from './background.js'
+import { retrieveWebsiteDetails, updateExtensionBadge, updateExtensionIcon } from './iconHandler.js'
+import { clearTabStates, getSimulationResults, removeTabState, setRpcConnectionStatus, updateTabState } from './storageVariables.js'
 import { setPrependTransactionsQueue } from '../simulation/services/SimulationModeEthereumClientService.js'
 import { Simulator } from '../simulation/simulator.js'
-import { TabConnection } from '../utils/user-interface-types.js'
+import { TabConnection, WebsiteTabConnections } from '../utils/user-interface-types.js'
 import { EthereumBlockHeader } from '../utils/wire-types.js'
 import { EthereumClientService } from '../simulation/services/EthereumClientService.js'
-import { sendPopupMessageToOpenWindows } from './backgroundUtils.js'
+import { getActiveAddress, getSocketFromPort, sendPopupMessageToOpenWindows, websiteSocketToString } from './backgroundUtils.js'
 import { sendSubscriptionMessagesForNewBlock } from '../simulation/services/EthereumSubscriptionService.js'
 import { refreshSimulation } from './popupMessageHandlers.js'
 import { isFailedToFetchError } from '../utils/errors.js'
-
-let simulator: Simulator | undefined = undefined
+import { Semaphore } from '../utils/semaphore.js'
+import { RawInterceptedRequest } from '../utils/requests.js'
+import { verifyAccess } from './accessManagement.js'
+import { InpageScriptRequest, TabState } from '../utils/interceptor-messages.js'
+import { replyToInterceptedRequest } from './messageSending.js'
+import { assertNever } from '../utils/typescript.js'
+import { ICON_NOT_ACTIVE } from '../utils/constants.js'
+import { connectedToSigner, ethAccountsReply, signerChainChanged, walletSwitchEthereumChainReply } from './providerMessageHandlers.js'
 
 const websiteTabConnections = new Map<number, TabConnection>()
 
-browser.runtime.onConnect.addListener(port => onContentScriptConnected(simulator, port, websiteTabConnections).catch(console.error))
 browser.tabs.onRemoved.addListener((tabId: number) => removeTabState(tabId))
 
 if (browser.runtime.getManifest().manifest_version === 2) {
 	clearTabStates()
 }
 
-async function newBlockAttemptCallback(blockheader: EthereumBlockHeader, ethereumClientService: EthereumClientService, isNewBlock: boolean) {
+function getProviderHandler(method: string) {
+	switch (method) {
+		case 'eth_accounts_reply': return { method: 'eth_accounts_reply' as const, func: ethAccountsReply }
+		case 'signer_chainChanged': return { method: 'signer_chainChanged' as const, func: signerChainChanged }
+		case 'wallet_switchEthereumChain_reply': return { method: 'wallet_switchEthereumChain_reply' as const, func: walletSwitchEthereumChainReply }
+		case 'connected_to_signer': return { method: 'connected_to_signer' as const, func: connectedToSigner }
+		default: return { method: 'notProviderMethod' as const }
+	}
+}
+
+export async function onContentScriptConnected(simulator: Simulator | undefined, port: browser.runtime.Port, websiteTabConnections: WebsiteTabConnections) {
+	const socket = getSocketFromPort(port)
+	if (port?.sender?.url === undefined) return
+	const websiteOrigin = (new URL(port.sender.url)).hostname
+	const websitePromise = retrieveWebsiteDetails(port, websiteOrigin)
+	const identifier = websiteSocketToString(socket)
+
+	console.log(`content script connected ${ websiteOrigin }`)
+
+	const tabConnection = websiteTabConnections.get(socket.tabId)
+	const newConnection = {
+		port: port,
+		socket: socket,
+		websiteOrigin: websiteOrigin,
+		approved: false,
+		wantsToConnect: false,
+	}
+	port.onDisconnect.addListener(() => {
+		const tabConnection = websiteTabConnections.get(socket.tabId)
+		if (tabConnection === undefined) return
+		delete tabConnection.connections[websiteSocketToString(socket)]
+		if (Object.keys(tabConnection).length === 0) {
+			websiteTabConnections.delete(socket.tabId)
+		}
+	})
+
+	const pendingRequestLimiter = new Semaphore(20) // only allow 20 requests pending at the time for a port
+
+	port.onMessage.addListener(async (payload) => {
+		if (!(
+			'data' in payload
+			&& typeof payload.data === 'object'
+			&& payload.data !== null
+			&& 'interceptorRequest' in payload.data
+		)) return
+		await pendingRequestLimiter.execute(async () => {
+			const rawMessage = RawInterceptedRequest.parse(payload.data)
+			const request = {
+				method: rawMessage.method,
+				...'params' in rawMessage ? { params: rawMessage.params } : {},
+				interceptorRequest: rawMessage.interceptorRequest,
+				usingInterceptorWithoutSigner: rawMessage.usingInterceptorWithoutSigner,
+				uniqueRequestIdentifier: { requestId: rawMessage.requestId, requestSocket: socket },
+			}
+			const activeAddress = await getActiveAddress(await getSettings(), socket.tabId)
+			const access = verifyAccess(websiteTabConnections, socket, request.method === 'eth_requestAccounts', websiteOrigin, activeAddress, await getSettings())
+			const providerHandler = getProviderHandler(request.method)
+			const identifiedMethod = providerHandler.method
+			if (identifiedMethod !== 'notProviderMethod') {
+				await providerHandler.func(simulator, websiteTabConnections, port, request, access)
+				const message: InpageScriptRequest = {
+					uniqueRequestIdentifier: request.uniqueRequestIdentifier,
+					method: identifiedMethod,
+					result: '0x' as const,
+				}
+				return replyToInterceptedRequest(websiteTabConnections, message)
+			}
+			if (access === 'noAccess' || activeAddress === undefined) {
+				if (request.method === 'eth_accounts') {
+					return replyToInterceptedRequest(websiteTabConnections, { method: 'eth_accounts' as const, result: [], uniqueRequestIdentifier: request.uniqueRequestIdentifier })
+				}
+				// if user has not given access, assume we are on chain 1
+				if (request.method === 'eth_chainId' || request.method === 'net_version') {
+					return replyToInterceptedRequest(websiteTabConnections, { method: 'eth_chainId' as const, result: 1n, uniqueRequestIdentifier: request.uniqueRequestIdentifier })
+				}
+			}
+			if (activeAddress === undefined) return refuseAccess(websiteTabConnections, request)
+
+			switch (access) {
+				case 'noAccess': return refuseAccess(websiteTabConnections, request)
+				case 'askAccess': return await gateKeepRequestBehindAccessDialog(simulator, websiteTabConnections, socket, request, await websitePromise, activeAddress, await getSettings())
+				case 'hasAccess': return await handleContentScriptMessage(simulator, websiteTabConnections, request, await websitePromise, activeAddress)
+				default: assertNever(access)
+			}
+		})
+	})
+
+	if (tabConnection === undefined) {
+		websiteTabConnections.set(socket.tabId, {
+			connections: { [identifier]: newConnection },
+		})
+		await updateTabState(socket.tabId, (previousState: TabState) => {
+			return {
+				...previousState,
+				tabIconDetails: {
+					icon: ICON_NOT_ACTIVE,
+					iconReason: 'No active address selected.',
+				}
+			}
+		})
+		updateExtensionIcon(websiteTabConnections, socket, websiteOrigin)
+	} else {
+		tabConnection.connections[identifier] = newConnection
+	}
+}
+
+async function newBlockAttemptCallback(blockheader: EthereumBlockHeader, ethereumClientService: EthereumClientService, isNewBlock: boolean, simulator: Simulator) {
 	const rpcConnectionStatus = {
 		isConnected: true,
 		lastConnnectionAttempt: new Date(),
@@ -57,9 +168,9 @@ async function onErrorBlockCallback(ethereumClientService: EthereumClientService
 
 async function startup() {
 	const settings = await getSettings()
-	if (settings.rpcNetwork.httpsRpc !== undefined) {
-		simulator = new Simulator(settings.rpcNetwork, newBlockAttemptCallback, onErrorBlockCallback)
-	}
+	if (settings.rpcNetwork.httpsRpc === undefined) throw new Error('RPC not set')
+	const simulator = new Simulator(settings.rpcNetwork, newBlockAttemptCallback, onErrorBlockCallback)
+	browser.runtime.onConnect.addListener(port => onContentScriptConnected(simulator, port, websiteTabConnections).catch(console.error))
 	if (simulator === undefined) throw new Error('simulator not found')
 	browser.runtime.onMessage.addListener(async function (message: unknown) {
 		if (simulator === undefined) throw new Error('Interceptor not ready yet')
