+++ app/ts/simulation/simulator.ts
@@ -38,8 +38,12 @@ const logHandler = new Map<string, Loghandler >([
 export class Simulator {
 	public ethereum: EthereumClientService
 
-	public constructor(rpcNetwork: RpcNetwork, newBlockAttemptCallback: (blockHeader: EthereumBlockHeader, ethereumClientService: EthereumClientService, isNewBlock: boolean) => void, onErrorBlockCallback: (ethereumClientService: EthereumClientService, error: Error) => void) {
-		this.ethereum = new EthereumClientService(new EthereumJSONRpcRequestHandler(rpcNetwork), newBlockAttemptCallback, onErrorBlockCallback)
+	public constructor(rpcNetwork: RpcNetwork, newBlockAttemptCallback: (blockHeader: EthereumBlockHeader, ethereumClientService: EthereumClientService, isNewBlock: boolean, simulator: Simulator) => void, onErrorBlockCallback: (ethereumClientService: EthereumClientService, error: Error) => void) {
+		this.ethereum = new EthereumClientService(
+			new EthereumJSONRpcRequestHandler(rpcNetwork),
+			(blockHeader: EthereumBlockHeader, ethereumClientService: EthereumClientService, isNewBlock: boolean) => newBlockAttemptCallback(blockHeader, ethereumClientService, isNewBlock, this),
+			onErrorBlockCallback
+		)
 	}
 
 	public cleanup = () => this.ethereum.cleanup()
